backend/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
python-multipart
sentence-transformers
pypdf
groq
python-dotenv